IBX-10667: Merge ThreeState and Checkbox components
