+++ packages/assets/src/scss/styles.scss
@@ -23,6 +23,5 @@
 @use 'inputs/checkbox';
 @use 'inputs/input-text';
 @use 'inputs/radio-button';
-@use 'inputs/three-state-checkbox';
 
 @use 'ui/clear-btn';
