+++ packages/components/src/components/Checkbox/CheckboxInput/CheckboxInput.tsx
@@ -6,13 +6,28 @@ import withStateChecked from '@ids-hoc/withStateChecked';
 
 import { CheckboxInputProps } from './CheckboxInput.types';
 
-export const CheckboxInput = ({ className = '', ...restProps }: CheckboxInputProps) => {
+export const CheckboxInput = ({ className = '', indeterminate = false, ...restProps }: CheckboxInputProps) => {
     const checkboxClassName = createCssClassNames({
         'ids-checkbox': true,
         [className]: true,
     });
+    const inputClassName = createCssClassNames({
+        'ids-input--indeterminate': indeterminate,
+    });
 
-    return <BaseChoiceInput {...restProps} className={checkboxClassName} type="checkbox" />;
+    return (
+        <BaseChoiceInput
+            className={checkboxClassName}
+            inputClassName={inputClassName}
+            ref={(node) => {
+                if (node) {
+                    node.indeterminate = indeterminate; // eslint-disable-line no-param-reassign
+                }
+            }}
+            type="checkbox"
+            {...restProps}
+        />
+    );
 };
 
 export const CheckboxInputStateful = withStateChecked<CheckboxInputProps>(CheckboxInput);
