+++ packages/components/src/components/Checkbox/CheckboxInput/CheckboxInput.types.ts
@@ -1,5 +1,7 @@
 import { BaseChoiceInputProps } from '@ids-partials/BaseChoiceInput';
 
 export type CheckboxInputProps = Omit<BaseChoiceInputProps, 'type'> & {
+    checked?: boolean;
+    indeterminate?: boolean;
     value: string;
 };
