--- packages/components/src/components/Checkbox/ThreeStateCheckboxInput/index.ts
@@ -1,2 +1,0 @@
-export * from './ThreeStateCheckboxInput';
-export * from './ThreeStateCheckboxInput.types';
