+++ packages/components/src/components/Checkbox/index.ts
@@ -1,4 +1,3 @@
 export * from './CheckboxField';
 export * from './CheckboxInput';
 export * from './CheckboxesListField';
-export * from './ThreeStateCheckboxInput';
